fix: refactor boardlist
