+++ src/page/BoardList/BoardListPage.jsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react";
-import { Card, Table } from "react-bootstrap";
-import { LinkContainer } from "react-router-bootstrap";
+import { Card } from "react-bootstrap";
 import { useNavigate } from "react-router-dom";
 import Pagination from "../../component/Pagination/Pagination";
 import { getBoardList } from "../../api";
+import BoardList from "../../component/BoardList/BoardList";
 
 const LOAD_STATUS = { loading: "loading", idle: "idle", error: "error" };
 
@@ -22,16 +22,8 @@ export default function BoardListPage() {
         setPageData(newPageData);
         setLoadStatus(LOAD_STATUS.idle);
       })
-      .catch((e) => {
-        console.log(e);
-        setLoadStatus(LOAD_STATUS.error);
-      });
+      .catch(() => setLoadStatus(LOAD_STATUS.error));
   }, [currentPageNumber]);
-
-  function formatDate(date) {
-    const result = `${date[0]}-${date[1]}-${date[2]} ${date[3]}:${date[4]}:${date[5]}`;
-    return result;
-  }
 
   function movePage(pageNumber) {
     setCurrentPageNumber(pageNumber);
@@ -55,30 +47,7 @@ export default function BoardListPage() {
 
   return (
     <>
-      <Table hover>
-        <thead>
-          <tr>
-            <th width="5%">번호</th>
-            <th width="40%">제목</th>
-            <th width="15%">작성자</th>
-            <th width="20%">날짜</th>
-            <th width="15%">조회수</th>
-          </tr>
-        </thead>
-        <tbody>
-          {boardList.map((board) => (
-            <LinkContainer key={board.id} to={`/board/${board.id}`} style={{ cursor: "pointer" }}>
-              <tr>
-                <td>{board.id}</td>
-                <td>{board.title}</td>
-                <td>{board.nickname}</td>
-                <td>{formatDate(board.registerDate)}</td>
-                <td>{board.hit}</td>
-              </tr>
-            </LinkContainer>
-          ))}
-        </tbody>
-      </Table>
+      <BoardList boardList={boardList} />
       <Pagination
         pageList={pageData.pageList}
         currentPageNumber={pageData.currentPageNumber}
